requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
astrapy==0.7.3
asttokens==2.4.1
attrs==23.2.0
cassandra-driver==3.29.0
cassio==0.1.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
executing==2.0.1
frozenlist==1.4.1
fsspec==2023.12.2
geomet==0.2.1.post1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
ipykernel==6.29.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
llama-index==0.9.42.post2
llama-parse==0.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.11.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pypdf==4.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
setuptools==69.0.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
